+++ src/test/java/org/jenkinsci/plugins/rolestrategy/playwright/RoleEndToEndUITest.java
@@ -7,6 +7,9 @@ import com.microsoft.playwright.Page;
 import com.microsoft.playwright.junit.UsePlaywright;
 import com.microsoft.playwright.options.AriaRole;
 import hudson.model.FreeStyleProject;
+import org.jenkinsci.plugins.rolestrategy.playwright.config.PlaywrightConfig;
+import org.jenkinsci.plugins.rolestrategy.playwright.helpers.UITestHelper;
+import org.jenkinsci.plugins.rolestrategy.playwright.pages.ManageAssignmentsPage;
 import org.junit.jupiter.api.Test;
 import org.jvnet.hudson.test.JenkinsRule;
 import org.jvnet.hudson.test.junit.jupiter.WithJenkins;
